requirements.txt
langchain
langchain_community
langchain_huggingface
sentence-transformers
faiss-cpu
pypdf
huggingface_hub
flask
flask-cors
python-dotenv
